+++ src/index/index.ts
@@ -1,0 +1,4 @@
+export * from './bruteForceIndex.js'
+export * from './fuseIndex.js'
+export * from './jsSearchIndex.js'
+export * from './types.js'